Add run_resource example group to test resource execution
This is an import of code rspec-puppet-augeas for selecting an individual
resource from the catalogue and actually executing it inside the spec to
ensure it runs successfully and in an idempotent manner.

This works well for providers based on file editing, and some utilities are
added to copy fixtures into a temporary directory (from the RSpec setting
resource_fixtures) for the provider to operate on in confinement.